+++ packages/cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@marklet/cli",
-  "version": "2.0.0",
+  "version": "2.0.1",
   "description": "A command line interface for marklet.",
   "author": "jjyyxx <1449843302@qq.com>",
   "contributors": [
@@ -19,7 +19,7 @@
     "url": "https://github.com/obstudio/Marklet/issues"
   },
   "dependencies": {
-    "@marklet/dev-server": "^1.1.1",
+    "@marklet/dev-server": "^1.1.2",
     "@marklet/parser": "^1.5.1",
     "chalk": "^2.4.1",
     "commander": "^2.18.0",
