+++ packages/dev-server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@marklet/dev-server",
-  "version": "1.1.1",
+  "version": "1.1.2",
   "description": "A develop server for marklet.",
   "author": "jjyyxx <1449843302@qq.com>",
   "contributors": [
