+++ packages/marklet/package.json
@@ -1,6 +1,6 @@
 {
   "name": "markletjs",
-  "version": "1.2.3",
+  "version": "1.3.0",
   "description": "A markup language designed for API manual pages.",
   "author": "jjyyxx <1449843302@qq.com>",
   "contributors": [
@@ -28,7 +28,7 @@
     "url": "https://github.com/obstudio/Marklet/issues"
   },
   "dependencies": {
-    "@marklet/cli": "^2.0.0",
+    "@marklet/cli": "^2.0.1",
     "@marklet/parser": "^1.5.1",
     "@marklet/renderer": "^1.3.1"
   }
